Added tests for rebases and moved older tests
