+++ gitgud/skills/test_levelbuilder.py
@@ -1,34 +1,8 @@
 import os
 
-from gitgud.skills import all_skills
+from gitgud.skills import all_skills, all_levels
 from gitgud.skills.util import NamedList, Skill
 from gitgud.skills.level_builder import Level
-
-
-def test_skill_access():
-    all_skills['1']
-    all_skills['basics']
-    all_skills['basics']['1']
-    all_skills['basics']['committing']
-
-
-def test_types():
-    for skill in all_skills:
-        assert isinstance(skill, Skill)
-        for level in skill:
-            assert isinstance(level, Level)
-
-
-def test_instructions():
-    for skill in all_skills:
-        for level in skill:
-            os.path.isfile(level.instructions_path)
-
-
-def test_goal():
-    for skill in all_skills:
-        for level in skill:
-            os.path.isfile(level.goal_path)
 
 
 def test_init_NL():
